+++ mex/editor/rules/state.py
@@ -7,6 +7,7 @@ from requests import HTTPError
 from starlette import status
 
 from mex.common.backend_api.connector import BackendApiConnector
+from mex.common.merged.main import create_merged_item
 from mex.common.models import (
     RULE_SET_REQUEST_CLASSES,
     RULE_SET_REQUEST_CLASSES_BY_NAME,
@@ -16,6 +17,7 @@ from mex.common.models import (
     AnyRuleSetResponse,
 )
 from mex.common.transform import ensure_postfix
+from mex.common.types import Identifier, Validation
 from mex.editor.exceptions import escalate_error
 from mex.editor.models import EditorValue
 from mex.editor.rules.models import EditorField, EditorPrimarySource, ValidationMessage
@@ -25,7 +27,10 @@ from mex.editor.rules.transform import (
     transform_validation_error_to_messages,
 )
 from mex.editor.state import State
-from mex.editor.transform import transform_models_to_stem_type
+from mex.editor.transform import (
+    transform_models_to_stem_type,
+    transform_models_to_title,
+)
 from mex.editor.utils import resolve_editor_value, resolve_identifier
 
 
@@ -33,9 +38,11 @@ class RuleState(State):
     """Base state for the edit and create components."""
 
     item_id: str | None = None
+    item_title: list[EditorValue] = []
     fields: list[EditorField] = []
     stem_type: str | None = None
     validation_messages: list[ValidationMessage] = []
+    has_changes: bool = False
 
     @rx.event(background=True)
     async def resolve_identifiers(self) -> None:
@@ -109,6 +116,14 @@ class RuleState(State):
             return
         if rule_set:
             self.stem_type = transform_models_to_stem_type([rule_set.additive])
+        if self.item_id:
+            preview = create_merged_item(
+                identifier=Identifier(self.item_id),
+                extracted_items=extracted_items,
+                rule_set=rule_set,
+                validation=Validation.LENIENT,
+            )
+            self.item_title = transform_models_to_title([preview])
         self.fields = transform_models_to_fields(
             extracted_items,
             additive=rule_set.additive,
@@ -142,6 +157,8 @@ class RuleState(State):
                 "backend", "error submitting rule set", exc.response.text
             )
             return
+
+        yield self.set_has_changes(False)
         # clear cache to show edits in the UI
         resolve_identifier.cache_clear()
         # trigger redirect to edit page or refresh state
@@ -225,40 +242,64 @@ class RuleState(State):
             index
         ].being_edited = not primary_source.editor_values[index].being_edited
 
+    def update_has_changes(self) -> EventSpec:
+        """Update the has changes value on client side."""
+        return rx.call_script(
+            f"window.updateMexEditorChanges({str(self.has_changes).lower()})",
+        )
+
     @rx.event
-    def add_additive_value(self, field_name: str) -> None:
+    def set_has_changes(self, value: bool) -> EventSpec:  # noqa: FBT001
+        """Set the has changes attribute to the given value.
+
+        Args:
+            value (bool): The value of the has changes attribute.
+        """
+        self.has_changes = value
+        return self.update_has_changes()
+
+    @rx.event
+    def add_additive_value(self, field_name: str) -> EventSpec:
         """Add an additive rule to the given field."""
         primary_source = self._get_editable_primary_source_by_field_name(field_name)
         primary_source.editor_values.append(EditorValue(being_edited=True))
+        return self.set_has_changes(True)
 
     @rx.event
-    def remove_additive_value(self, field_name: str, index: int) -> None:
+    def remove_additive_value(self, field_name: str, index: int) -> EventSpec:
         """Remove an additive rule from the given field."""
         primary_source = self._get_editable_primary_source_by_field_name(field_name)
         primary_source.editor_values.pop(index)
+        return self.set_has_changes(True)
 
     @rx.event
-    def set_text_value(self, field_name: str, index: int, value: str) -> None:
+    def set_text_value(self, field_name: str, index: int, value: str) -> EventSpec:
         """Set the text attribute on an additive editor value."""
         primary_source = self._get_editable_primary_source_by_field_name(field_name)
         primary_source.editor_values[index].text = value
+        return self.set_has_changes(True)
 
     @rx.event
-    def set_identifier_value(self, field_name: str, index: int, value: str) -> None:
+    def set_identifier_value(
+        self, field_name: str, index: int, value: str
+    ) -> EventSpec:
         """Set the identifier attribute on an additive editor value."""
         primary_source = self._get_editable_primary_source_by_field_name(field_name)
         primary_source.editor_values[index].identifier = value
         primary_source.editor_values[index].href = f"/item/{value}"
+        return self.set_has_changes(True)
 
     @rx.event
-    def set_badge_value(self, field_name: str, index: int, value: str) -> None:
+    def set_badge_value(self, field_name: str, index: int, value: str) -> EventSpec:
         """Set the badge attribute on an additive editor value."""
         primary_source = self._get_editable_primary_source_by_field_name(field_name)
         primary_source.editor_values[index].badge = value
+        return self.set_has_changes(True)
 
     @rx.event
-    def set_href_value(self, field_name: str, index: int, value: str) -> None:
+    def set_href_value(self, field_name: str, index: int, value: str) -> EventSpec:
         """Set an external href on an additive editor value."""
         primary_source = self._get_editable_primary_source_by_field_name(field_name)
         primary_source.editor_values[index].href = value
         primary_source.editor_values[index].external = True
+        return self.set_has_changes(True)